requirements.txt
Django==1.9.8
djangorestframework==3.3.3
django-rest-swagger==2.1.2
